move file helper to util package
